+++ packages/build-webpack-config/src/index.ts
@@ -1,6 +1,6 @@
 const applyCliOption = require('./applyCliOption');
 const applyUserConfig = require('./applyUserConfig');
-const getWebpackBase = require('./base');
+const getWebpackBase = require('./getWebpackBase');
 const getDefaultConfig = require('./config/default.config');
 
 export {
